refactor(storage_impl): Move Conversion impl for customer, payment_methods to storage_impl crate
