+++ crates/router/src/core/payouts.rs
@@ -48,8 +48,6 @@ use time::Duration;
 #[cfg(all(feature = "olap", feature = "payouts"))]
 use crate::consts as payout_consts;
 #[cfg(feature = "olap")]
-use crate::types::domain::behaviour::Conversion;
-#[cfg(feature = "olap")]
 use crate::types::PayoutActionData;
 use crate::{
     core::{
@@ -996,7 +994,7 @@ pub async fn payouts_filtered_list_core(
         join_all(list.into_iter().map(|(p, pa, customer, address)| async {
             let customer: Option<domain::Customer> = customer
                 .async_and_then(|cust| async {
-                    domain::Customer::convert_back(
+                    <domain::Customer as storage_impl::behaviour::Conversion>::convert_back(
                         &(&state).into(),
                         cust,
                         &(platform.get_processor().get_key_store().clone()).key,
@@ -1018,7 +1016,7 @@ pub async fn payouts_filtered_list_core(
 
             let payout_addr: Option<payment_enums::Address> = address
                 .async_and_then(|addr| async {
-                    domain::Address::convert_back(
+                    <domain::Address as domain::behaviour::Conversion>::convert_back(
                         &(&state).into(),
                         addr,
                         &(platform.get_processor().get_key_store().clone()).key,
